% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-figtree.R
\docType{data}
\name{font_figtree}
\alias{font_figtree}
\title{Figtree font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_figtree
}
\description{
\code{font_figtree} == "\code{Figtree}"
}
\keyword{datasets}
