% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-figtree.R
\name{import_figtree}
\alias{import_figtree}
\title{Import Figtree font for use in charts}
\usage{
import_figtree()
}
\description{
Import Figtree font for use in charts
}
\note{
The location of the font directory is displayed. It is highly
recommended that you install them on your system the same way you would any
other font you wish to use in other programs.
}
