% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-install.R
\name{install_fonts_macos}
\alias{install_fonts_macos}
\title{Import fonts in the ktheme package user-wide or system-wide (on Mac OS)}
\usage{
install_fonts_macos(font_dst = "~/Library/Fonts")
}
\arguments{
\item{font_dst}{location for fonts, default is "~/Library/Fonts" for user wide installation, use "/Library/Fonts" for
system-wide (may require root privs)}
}
\description{
Import fonts in the ktheme package user-wide or system-wide (on Mac OS)
}
