% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{kth_b64_favicon_svg}
\alias{kth_b64_favicon_svg}
\title{KTH favicon}
\usage{
kth_b64_favicon_svg(
  logo_file = system.file(package = "ktheme", "logo", "favicon.svg"),
  as = c("img", "link", "uri"),
  ...
)
}
\arguments{
\item{logo_file}{optional path to custom logo}

\item{as}{one of "img" or "link"}

\item{...}{further arguments sent to HTML img if using as "img"}
}
\value{
HTML
}
\description{
Favicon in SVG format.
}
\details{
If no logo_file path is provided, the embedded favicon will be used.
}
