% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{kth_b64_logo_svg}
\alias{kth_b64_logo_svg}
\title{KTH logo}
\usage{
kth_b64_logo_svg(
  variant = c("blue", "black", "white"),
  as = c("img", "uri"),
  ...
)
}
\arguments{
\item{variant, }{one of "blue", "black" or "white"}

\item{as}{return type, one of "img" or "uri"}

\item{...}{further arguments for the HTML img tag}
}
\value{
HTML browsable image tag with base64 encoded data embedded or uri
}
\description{
Logo in SVG format.
}
