% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{kth_b64_webfont}
\alias{kth_b64_webfont}
\title{KTH webfont}
\usage{
kth_b64_webfont(
  font_name = c("figtree", "georgia", "open-sans", "open-sans-condensed"),
  as = c("css", "uri")
)
}
\arguments{
\item{font_name, }{one of "figtree", "georgia", "open-sans" or
"open-sans-condensed"}

\item{as}{return type, one of "css" or "uri"}
}
\value{
CSS tag with base64 encoded data embedded or uri
}
\description{
Webfont for CSS styling in woff2 format
}
