% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{kth_colors}
\alias{kth_colors}
\title{KTH Profile colors}
\usage{
kth_colors(color_name)
}
\arguments{
\item{color_name}{the color name(s), by default missing}
}
\value{
a vector of RGB hex strings for requested named colors
}
\description{
Get the full set of KTH Profile colors (RGB hex) or a specific named color
}
