% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{kth_pal}
\alias{kth_pal}
\title{A qualitative color palette with colors from KTH's graphical profile}
\usage{
kth_pal(pal = palette_kth_neo())
}
\arguments{
\item{pal}{the palette to use, by default palette_kth_neo()}
}
\description{
A qualitative color palette with colors from KTH's graphical profile
}
\examples{
library(scales)
scales::show_col(kth_pal()(5))
}
