% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktheme-package.R
\name{ktheme-exports}
\alias{ktheme-exports}
\alias{\%>\%}
\title{ktheme exported operators}
\description{
The following functions are imported and then re-exported
from the ktheme package to enable use of the magrittr
pipe operator with no additional library calls
}
