% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-open-sans-condensed.R
\docType{data}
\name{font_osc}
\alias{font_osc}
\title{Open Sans font name R variable aliases}
\format{length 1 character vector}
\usage{
font_osc
}
\description{
\code{font_osc} == "\verb{Open Sans Condensed}"
}
\keyword{datasets}
