% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_kth_info}
\alias{palette_kth_info}
\title{KTH color palette information for KTH standard color palette}
\usage{
palette_kth_info()
}
\value{
data frame with information about the available palette(s)
}
\description{
This function aligns its signature r RColorBrewer::color.pal.info()
}
