% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_kth_neo}
\alias{palette_kth_neo}
\title{KTH color palette with 5 qualitative colors}
\usage{
palette_kth_neo(n = 12, name = "KTH", type = c("qual", "seq", "div"))
}
\arguments{
\item{n}{number of colors to use (1..6 x 3), default is to return}

\item{name}{name of palette}

\item{type}{nature of data, default "qual" for qualitative or "seq" for
sequential or "div" for diverging}
}
\value{
named vector with colors and hex codes
}
\description{
This palette is taken from https://intra.kth.se/en/administration/kommunikation/grafiskprofil/profilfarger-1.845077
}
\details{
This function aligns its signature with RColorBrewer::color.pal()
}
\examples{
palette_kth_neo(12)  # return the primary color in the KTH palette
}
