% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{sdg_colors}
\alias{sdg_colors}
\title{SDG color table from UN}
\format{
A data frame with 17 rows and 4 variables:
\describe{
  \item{\code{goal}}{integer number for the SDG, 1-17}
  \item{\code{color}}{character RGB color}
  \item{\code{short_desc}}{character Short description of SDG}
  \item{\code{long_desc}}{character Long description of SDG}
}
}
\usage{
sdg_colors()
}
\description{
SDG color table from UN
}
