% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-kth.R
\name{theme_kth_figtree}
\alias{theme_kth_figtree}
\title{A [ggplot2] theme with default for the KTH graphical profile using the Figtree font .}
\usage{
theme_kth_figtree(
  base_family = "Figtree-Regular",
  plot_title_family = "Figtree",
  subtitle_face = "bold",
  strip_text_face = "bold",
  axis_title_face = "bold",
  fontscale_factor = 1.2,
  ...
)
}
\arguments{
\item{base_family}{base font family and size}

\item{plot_title_family}{plot title family}

\item{subtitle_face}{the font face used for subtitles, by default bold}

\item{strip_text_face}{the font face used for strip texts, by default bold}

\item{axis_title_face}{the font face used for axis titles, by default bold}

\item{fontscale_factor}{the font scaling factor used, by default 1.2}

\item{...}{
  Arguments passed on to \code{\link[=theme_kth]{theme_kth}}
  \describe{
    \item{\code{base_family,base_size}}{base font family and size}
    \item{\code{plot_title_family,plot_title_face,plot_title_size,plot_title_margin}}{plot title family, face, size and margi}
    \item{\code{subtitle_family,subtitle_face,subtitle_size}}{plot subtitle family, face and size}
    \item{\code{subtitle_margin}}{plot subtitle margin bottom (single numeric value)}
    \item{\code{strip_text_family,strip_text_face,strip_text_size}}{facet label font family, face and size}
    \item{\code{caption_family,caption_face,caption_size,caption_margin}}{plot caption family, face, size and margin}
    \item{\code{axis_title_family,axis_title_face,axis_title_size}}{axis title font family, face and size}
    \item{\code{axis_title_just}}{axis title font justification, one of `[blmcrt]`}
    \item{\code{plot_margin}}{plot margin (specify with [ggplot2::margin()])}
    \item{\code{grid_col,axis_col}}{grid & axis colors; both default to `#cccccc`}
    \item{\code{grid}}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}
    \item{\code{axis_text_size}}{font size of axis text}
    \item{\code{axis}}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}
    \item{\code{ticks}}{ticks if `TRUE` add ticks}
  }}
}
\description{
A [ggplot2] theme with default for the KTH graphical profile using the Figtree font .
}
