% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-kth.R
\name{theme_kth_neo}
\alias{theme_kth_neo}
\title{A \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} theme with default for the KTH graphical profile using the Figtree font .}
\usage{
theme_kth_neo(
  base_family = "Figtree Light",
  plot_title_family = "Figtree",
  subtitle_face = "plain",
  strip_text_face = "plain",
  axis_title_face = "plain",
  fontscale_factor = 1.1,
  ...
)
}
\arguments{
\item{base_family}{base font family and size}

\item{plot_title_family}{plot title family}

\item{subtitle_face}{the font face used for subtitles, by default plain}

\item{strip_text_face}{the font face used for strip texts, by default plain}

\item{axis_title_face}{the font face used for axis titles, by default plain}

\item{fontscale_factor}{the font scaling factor used, by default 1.0}

\item{...}{
  Arguments passed on to \code{\link[=theme_kth]{theme_kth}}
  \describe{
    \item{\code{base_family,base_size}}{base font family and size}
    \item{\code{plot_title_family,plot_title_face,plot_title_size,plot_title_margin}}{plot title family, face, size and margi}
    \item{\code{subtitle_family,subtitle_face,subtitle_size}}{plot subtitle family, face and size}
    \item{\code{subtitle_margin}}{plot subtitle margin bottom (single numeric value)}
    \item{\code{strip_text_family,strip_text_face,strip_text_size}}{facet label font family, face and size}
    \item{\code{caption_family,caption_face,caption_size,caption_margin}}{plot caption family, face, size and margin}
    \item{\code{axis_title_family,axis_title_face,axis_title_size}}{axis title font family, face and size}
    \item{\code{axis_title_just}}{axis title font justification, one of \verb{[blmcrt]}}
    \item{\code{plot_margin}}{plot margin (specify with \code{\link[ggplot2:element]{ggplot2::margin()}})}
    \item{\code{grid_col,axis_col}}{grid & axis colors; both default to \verb{#cccccc}}
    \item{\code{grid}}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of \code{X}, \code{x}, \code{Y}, \code{y})}
    \item{\code{axis_text_size}}{font size of axis text}
    \item{\code{axis}}{add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}"}
    \item{\code{ticks}}{ticks if \code{TRUE} add ticks}
  }}
}
\description{
A \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} theme with default for the KTH graphical profile using the Figtree font .
}
