% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-kth.R
\name{theme_kth_osc}
\alias{theme_kth_osc}
\title{A [ggplot2] theme with default for the KTH graphical profile using the Narrow or Condensed
version of Open Sans.}
\usage{
theme_kth_osc(
  base_family = "Open Sans Condensed",
  plot_title_family = "Open Sans Condensed",
  ...
)
}
\arguments{
\item{base_family}{base font family and size}

\item{plot_title_family}{plot title family}

\item{...}{
  Arguments passed on to \code{\link[=theme_kth]{theme_kth}}
  \describe{
    \item{\code{base_size}}{base font family and size}
    \item{\code{plot_title_face}}{plot title family, face, size and margi}
    \item{\code{plot_title_size}}{plot title family, face, size and margi}
    \item{\code{plot_title_margin}}{plot title family, face, size and margi}
    \item{\code{subtitle_family}}{plot subtitle family, face and size}
    \item{\code{subtitle_face}}{plot subtitle family, face and size}
    \item{\code{subtitle_size}}{plot subtitle family, face and size}
    \item{\code{subtitle_margin}}{plot subtitle margin bottom (single numeric value)}
    \item{\code{strip_text_family}}{facet label font family, face and size}
    \item{\code{strip_text_face}}{facet label font family, face and size}
    \item{\code{strip_text_size}}{facet label font family, face and size}
    \item{\code{caption_family}}{plot caption family, face, size and margin}
    \item{\code{caption_face}}{plot caption family, face, size and margin}
    \item{\code{caption_size}}{plot caption family, face, size and margin}
    \item{\code{caption_margin}}{plot caption family, face, size and margin}
    \item{\code{axis_title_family}}{axis title font family, face and size}
    \item{\code{axis_title_face}}{axis title font family, face and size}
    \item{\code{axis_title_size}}{axis title font family, face and size}
    \item{\code{axis_title_just}}{axis title font justification, one of `[blmcrt]`}
    \item{\code{plot_margin}}{plot margin (specify with [ggplot2::margin()])}
    \item{\code{grid_col}}{grid & axis colors; both default to `#cccccc`}
    \item{\code{axis_col}}{grid & axis colors; both default to `#cccccc`}
    \item{\code{grid}}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}
    \item{\code{axis_text_size}}{font size of axis text}
    \item{\code{axis}}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}
    \item{\code{ticks}}{ticks if `TRUE` add ticks}
  }}
}
\description{
A [ggplot2] theme with default for the KTH graphical profile using the Narrow or Condensed
version of Open Sans.
}
