% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{unpaywall_colors}
\alias{unpaywall_colors}
\title{A data frame with colors for OA statuses}
\usage{
unpaywall_colors()
}
\value{
A data frame with columns oa_status, oa_color
}
\description{
A data frame with colors for OA statuses
}
