% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto_extension.R
\name{use_quarto_ext}
\alias{use_quarto_ext}
\title{Function to open a KTH quarto template}
\usage{
use_quarto_ext(file_name = NULL, ext_name = "ktheme")
}
\arguments{
\item{file_name}{a directory name}

\item{ext_name}{an extension name, by default "ktheme"}
}
\description{
Function to open a KTH quarto template
}
